Add files via upload
The development of this C++ project has been an enriching experience for the team,
providing us with valuable insights into various interesting concepts on C++. After
careful consideration, we determined that Dev C++ was the ideal compiler for the
project, and we successfully built a large-scale project using it for the first time. The
process of learning and utilizing Dev C++ was enjoyable, and we gathered the required
header files from the internet before putting our classroom knowledge and concepts to
work on the quiz game.
The welcome page includes a switch statement that enables users to choose from
starting the game, reading the instructions, viewing past results, or exiting the game.
Once the user provides their information, the quiz commences with a 60-second ticking
counter. We incorporated the use of if else if statements to allow users to skip questions
or select from four different answer choices.
After the quiz is completed, the results are displayed through an animated bear, which
we created by adding various signs and symbols. Users are returned to the welcome
screen, where they can choose to play the quiz again, view previous results, or quit the
game. Additionally, we utilized a file to store user information and their quiz results for
future reference.
We faced certain limitations and challenges throughout the program writing and
analysis process, which prevented us from incorporating complex graphics and
providing users with topic options before beginning the quiz. However, despite these
issues, we consider the project a success.
This project served as an excellent learning opportunity for the team members, as we
applied theoretical knowledge into practical use. Collaborative problem-solving played
a significant role, and the use of various C++ concepts, such as classes, loops, arrays,
functions, and conditional statements, helped us create the quiz's functionality. The
project helped us gain knowledge about the development cycle, such as debugging,
building, and testing programs. This project has provided us with a strong foundation
for future projects and has taught us essential skills required for writing clean, efficient
and maintainable code. We anticipate more such opportunities to grow and enhance our
skills as a team.